+++ packages/cli/templates/angular/ig-ts/funnel-chart/default/files/src/app/components/__path__/__filePrefix__.component.ts
@@ -6,17 +6,17 @@ import { funnelData } from "../../assets/funnel-data";
 	template: `
 		<h1>$(description)</h1>
 		<div style="float: left">
-			<ig-funnel-chart [(options)]="normalOptions" widgetId='chartNormal'>
+			<ig-funnel-chart [(options)]="normalOptions" widgetId="chartNormal">
 			</ig-funnel-chart>
 		</div>
 
 		<div style="float: left; width: 500px; height: 500px;">
-			<ig-funnel-chart [(options)]="weightedOptions" widgetId='chartWeighted'>
+			<ig-funnel-chart [(options)]="weightedOptions" widgetId="chartWeighted">
 			</ig-funnel-chart>
 		</div>
 
 			<div style="float: left; width: 500px; height: 500px;">
-			<ig-funnel-chart [(options)]="invertedOptions" widgetId='chartInverted'>
+			<ig-funnel-chart [(options)]="invertedOptions" widgetId="chartInverted">
 			</ig-funnel-chart>
 
 		</div>