+++ packages/igx-templates/igx-ts/custom-templates/fintech-grid/index.ts
@@ -28,7 +28,7 @@ class IgxFinTechGridTemplate extends IgniteUIForAngularTemplate {
 			},
 			{ import: "FormsModule", from: "@angular/forms" }
 		];
-		this.packages = ["igniteui-angular-core@~10.1.3", "igniteui-angular-charts@~10.1.3"];
+		this.packages = ["igniteui-angular-core@~11.1.0", "igniteui-angular-charts@~11.1.0"];
 	}
 }
 module.exports = new IgxFinTechGridTemplate();